:art: FLOW-647 helper functions segregated
