+++ packages/@cldcvr/flow-form-builder/html.html-data.json
@@ -83,7 +83,7 @@
     },
     {
       "name": "f-form-builder",
-      "description": "Events:\n\n  * `input` {`CustomEvent<{ [x: string]: Record<string, unknown>; } | { [x: string]: Record<string, unknown>[]; }>`} - \n\n  * `submit` {`CustomEvent<FormBuilderObjectGroupValues | FormBuilderArrayGroupValues>`} - \n\n  * `stateChange` {`CustomEvent<{ isValid: boolean; isChanged: boolean; errors: Record<string, string>; refs: Record<string, Ref<FFormInputElements>>; helperTexts: Record<string, string | undefined>; rules: Record<...>; errorRefs: Record<...>; showFunctions: Map<...>; suffixFunctions?: Map<...> | undefined; }>`} - \n\nAttributes:\n\n  * `config` {`FormBuilderConfig`} - \n\n  * `values` - \n\n  * `tooltip` {`string | undefined`} - \n\nProperties:\n\n  * `config` {`FormBuilderConfig`} - \n\n  * `values` - \n\n  * `state` {`FormBuilderState`} - Internal state of formbuilder\n\n  * `formRef` - form reference\n\n  * `groups` {`InternalFormBuilderGroup[]`} - \n\n  * `removedGroupName` {`string | undefined`} - holds name of last deleted group\n\n  * `validateField` - \n\n  * `validateForm` - \n\n  * `bindValidation` - \n\n  * `renderFields` - \n\n  * `renderGroups` - \n\n  * `removeGroup` - \n\n  * `duplicateGroup` - \n\n  * `handleGroupDuplicate` - \n\n  * `styles` {`CSSResult[]`} - css loaded from scss file\n\n  * `tooltipElement` {`HTMLElement`} - \n\n  * `tooltip` {`string | undefined`} - ",
+      "description": "Events:\n\n  * `input` {`CustomEvent<{ [x: string]: Record<string, unknown>; } | { [x: string]: Record<string, unknown>[]; }>`} - \n\n  * `submit` {`CustomEvent<FormBuilderObjectGroupValues | FormBuilderArrayGroupValues>`} - \n\n  * `stateChange` {`CustomEvent<{ isValid: boolean; isChanged: boolean; errors: Record<string, string>; refs: Record<string, Ref<FFormInputElements>>; helperTexts: Record<string, string | undefined>; rules: Record<...>; errorRefs: Record<...>; showFunctions: Map<...>; suffixFunctions?: Map<...> | undefined; }>`} - \n\nAttributes:\n\n  * `config` {`FormBuilderConfig`} - \n\n  * `values` - \n\n  * `tooltip` {`string | undefined`} - \n\nProperties:\n\n  * `config` {`FormBuilderConfig`} - \n\n  * `values` - \n\n  * `state` {`FormBuilderState`} - Internal state of formbuilder\n\n  * `formRef` - form reference\n\n  * `groups` {`InternalFormBuilderGroup[]`} - \n\n  * `removedGroupName` {`string | undefined`} - holds name of last deleted group\n\n  * `validateField` - \n\n  * `validateForm` - \n\n  * `bindValidation` - \n\n  * `renderFields` - \n\n  * `renderGroups` - \n\n  * `removeGroup` - \n\n  * `duplicateGroup` - \n\n  * `handleGroupDuplicate` - \n\n  * `checkSubmit` - \n\n  * `checkFieldType` - \n\n  * `checkSuffixConditions` - \n\n  * `checkAllShowConditions` - \n\n  * `styles` {`CSSResult[]`} - css loaded from scss file\n\n  * `tooltipElement` {`HTMLElement`} - \n\n  * `tooltip` {`string | undefined`} - ",
       "attributes": [
         {
           "name": "config",
